Populated queries.py in blueprints
